reset routing
