+++ src/components/NavBar.vue
@@ -47,7 +47,8 @@ export default {
       return (
         this.$route.name === 'welcome' ||
         this.$route.name === 'onboard-resources' ||
-        this.$route.name === 'name-team'
+        this.$route.name === 'name-team' ||
+        this.$route.name === 'accept'
       )
     }
   },
