+++ src/router.js
@@ -36,7 +36,7 @@ export const routes = [
     path: '/accept',
     component: () =>
       import(
-        /* webpackChunkName: "accept" */ '@/pages/AcceptInvitationPage.vue'
+        /* webpackChunkName: "accept" */ '@/pages/Onboard/AcceptInvitationPage.vue'
       ),
     beforeEnter: multiguard([authNavGuard, tenantNavGuard])
   },
